requirements.txt
GitPython
markdown2
huggingface_hub
langchain
langchain_community